+++ public/js/index.js
@@ -1096,8 +1096,8 @@ function () {
       return data;
     }
   }, {
-    key: "exportImage",
-    value: function exportImage() {
+    key: "exportImageAsDataURL",
+    value: function exportImageAsDataURL() {
       var offscreen = document.createElement('canvas');
       offscreen.setAttribute('height', this.canvasArea.style.height);
       offscreen.setAttribute('width', this.canvasArea.style.width);
@@ -1105,6 +1105,18 @@ function () {
       offscreenContext.drawImage(this.background, 0, 0);
       offscreenContext.drawImage(this.canvas, 0, 0);
       return offscreen.toDataURL("image/png");
+    }
+  }, {
+    key: "exportPathAsSaveable",
+    value: function exportPathAsSaveable() {
+      var win = window.open();
+      win.document.write('<iframe src="' + this.canvas.toDataURL('image/png') + '" frameborder="0" style="background:#f1f1f1; border:0; top:0px; left:0px; bottom:0px; right:0px; width:100%; height:100vh;" allowfullscreen></iframe>');
+    }
+  }, {
+    key: "exportImageAsSaveable",
+    value: function exportImageAsSaveable() {
+      var win = window.open();
+      win.document.write('<iframe src="' + this.exportImageAsDataURL() + '" frameborder="0" style="border:0; top:0px; left:0px; bottom:0px; right:0px; width:100%; height:100vh;" allowfullscreen></iframe>');
     }
   }, {
     key: "export",