+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [develop]
+### Added
+- Makes images saveable
+
 ### Changed
 - Replaces webpack
 - Implements sass
-- Allows scaling down a large canvas
 
 ## [1.1.0]
 ### Added